Support 64-bit Linux, misc bugfixes

+++ macip.h
@@ -24,7 +24,9 @@
  * SUCH DAMAGE.
  */
 
-extern int  macip_open (char *zone, u_long net, u_long mask, u_long ns, outputfunc_t o);
+#include <stdint.h>
+
+extern int  macip_open (char *zone, uint32_t net, uint32_t mask, uint32_t ns, outputfunc_t o);
 extern void macip_close (void);
 extern void macip_idle (void);
